Add ability to extend or truncate target arrays
When applying values to each element of an array and the array of values
has a different number of elements than the target array.

Also make applying a value to every element happen after applying
a value to each element, so that arrays that are extended get the 'every'
value set on their new elements as well as their existing elements.